clean up utils.go
